chore(space-wizard): set default configurator
